+++ agent-inject/agent/annotations.go
@@ -37,6 +37,14 @@ const (
 	// secret-volume-path annotation.
 	AnnotationAgentInjectFile = "vault.hashicorp.com/agent-inject-file"
 
+	// AnnotationAgentInjectFilePermission is the key of the annotation that contains the
+	// permission of the file to create on disk. The name of the
+	// secret is the string after "vault.hashicorp.com/agent-inject-perms-", and
+	// should map to the same unique value provided in
+	// "vault.hashicorp.com/agent-inject-secret-". The value is the value of the permission, for
+	// example "0644"
+	AnnotationAgentInjectFilePermission = "vault.hashicorp.com/agent-inject-perms"
+
 	// AnnotationAgentInjectTemplate is the key annotation that configures Vault
 	// Agent what template to use for rendering the secrets.  The name
 	// of the template is any unique string after "vault.hashicorp.com/agent-inject-template-",
@@ -231,26 +239,32 @@ const (
 	// will exit on template render failures once it has exhausted all its retry
 	// attempts. Defaults to true.
 	AnnotationTemplateConfigExitOnRetryFailure = "vault.hashicorp.com/template-config-exit-on-retry-failure"
+
+	// AnnotationTemplateConfigStaticSecretRenderInterval
+	// If specified, configures how often Vault Agent Template should render non-leased secrets such as KV v2.
+	// Defaults to 5 minutes.
+	AnnotationTemplateConfigStaticSecretRenderInterval = "vault.hashicorp.com/template-static-secret-render-interval"
 )
 
 type AgentConfig struct {
-	Image              string
-	Address            string
-	AuthType           string
-	AuthPath           string
-	Namespace          string
-	RevokeOnShutdown   bool
-	UserID             string
-	GroupID            string
-	SameID             bool
-	SetSecurityContext bool
-	ProxyAddress       string
-	DefaultTemplate    string
-	ResourceRequestCPU string
-	ResourceRequestMem string
-	ResourceLimitCPU   string
-	ResourceLimitMem   string
-	ExitOnRetryFailure bool
+	Image                      string
+	Address                    string
+	AuthType                   string
+	AuthPath                   string
+	Namespace                  string
+	RevokeOnShutdown           bool
+	UserID                     string
+	GroupID                    string
+	SameID                     bool
+	SetSecurityContext         bool
+	ProxyAddress               string
+	DefaultTemplate            string
+	ResourceRequestCPU         string
+	ResourceRequestMem         string
+	ResourceLimitCPU           string
+	ResourceLimitMem           string
+	ExitOnRetryFailure         bool
+	StaticSecretRenderInterval string
 }
 
 // Init configures the expected annotations required to create a new instance
@@ -402,6 +416,9 @@ func Init(pod *corev1.Pod, cfg AgentConfig) error {
 	if _, ok := pod.ObjectMeta.Annotations[AnnotationTemplateConfigExitOnRetryFailure]; !ok {
 		pod.ObjectMeta.Annotations[AnnotationTemplateConfigExitOnRetryFailure] = strconv.FormatBool(cfg.ExitOnRetryFailure)
 	}
+	if _, ok := pod.ObjectMeta.Annotations[AnnotationTemplateConfigStaticSecretRenderInterval]; !ok {
+		pod.ObjectMeta.Annotations[AnnotationTemplateConfigStaticSecretRenderInterval] = cfg.StaticSecretRenderInterval
+	}
 
 	return nil
 }
@@ -458,6 +475,11 @@ func (a *Agent) secrets() []*Secret {
 			file := fmt.Sprintf("%s-%s", AnnotationAgentInjectFile, raw)
 			if val, ok := a.Annotations[file]; ok {
 				s.FilePathAndName = val
+			}
+
+			filePerm := fmt.Sprintf("%s-%s", AnnotationAgentInjectFilePermission, raw)
+			if val, ok := a.Annotations[filePerm]; ok {
+				s.FilePermission = val
 			}
 
 			secrets = append(secrets, s)