+++ agent-inject/agent/config.go
@@ -77,6 +77,7 @@ type Template struct {
 	RightDelim     string `json:"right_delimiter,omitempty"`
 	Command        string `json:"command,omitempty"`
 	Source         string `json:"source,omitempty"`
+	Perms          string `json:"perms,omitempty"`
 }
 
 // Listener defines the configuration for Vault Agent Cache Listener
@@ -103,7 +104,8 @@ type CachePersist struct {
 
 // TemplateConfig defines the configuration for template_config in Vault Agent
 type TemplateConfig struct {
-	ExitOnRetryFailure bool `json:"exit_on_retry_failure"`
+	ExitOnRetryFailure         bool   `json:"exit_on_retry_failure"`
+	StaticSecretRenderInterval string `json:"static_secret_render_interval,omitempty"`
 }
 
 func (a *Agent) newTemplateConfigs() []*Template {
@@ -135,6 +137,9 @@ func (a *Agent) newTemplateConfigs() []*Template {
 			LeftDelim:   "{{",
 			RightDelim:  "}}",
 			Command:     secret.Command,
+		}
+		if secret.FilePermission != "" {
+			tmpl.Perms = secret.FilePermission
 		}
 		templates = append(templates, tmpl)
 	}
@@ -173,6 +178,7 @@ func (a *Agent) newConfig(init bool) ([]byte, error) {
 		Templates: a.newTemplateConfigs(),
 		TemplateConfig: &TemplateConfig{
 			ExitOnRetryFailure: a.VaultAgentTemplateConfig.ExitOnRetryFailure,
+			StaticSecretRenderInterval: a.VaultAgentTemplateConfig.StaticSecretRenderInterval,
 		},
 	}