+++ agent-inject/agent/config_test.go
@@ -224,6 +224,9 @@ func TestFilePathAndName(t *testing.T) {
 			}
 
 			agent, err := New(pod, patches)
+			if err != nil {
+				t.Errorf("got error creating agent, shouldn't have: %s", err)
+			}
 			cfg, err := agent.newConfig(true)
 			if err != nil {
 				t.Errorf("got error creating Vault config, shouldn't have: %s", err)
@@ -235,6 +238,95 @@ func TestFilePathAndName(t *testing.T) {
 			}
 			if config.Templates[0].Destination != tt.destination {
 				t.Errorf("wrong destination: %s != %s", config.Templates[0].Destination, tt.destination)
+			}
+		})
+	}
+}
+
+func TestFilePermission(t *testing.T) {
+
+	tests := []struct {
+		name        string
+		annotations map[string]string
+		permission  string
+	}{
+		{
+			"just secret",
+			map[string]string{
+				"vault.hashicorp.com/agent-inject-secret-foo": "db/creds/foo",
+				"vault.hashicorp.com/agent-inject-perms-foo":  "0600",
+			},
+			"0600",
+		},
+		{
+			"just secret without permission",
+			map[string]string{
+				"vault.hashicorp.com/agent-inject-secret-foo": "db/creds/foo",
+			},
+			"",
+		},
+		{
+			"with relative file path",
+			map[string]string{
+				"vault.hashicorp.com/agent-inject-secret-foo": "db/creds/foo",
+				"vault.hashicorp.com/agent-inject-file-foo":   "nested/foofile",
+				"vault.hashicorp.com/agent-inject-perms-foo":  "0600",
+			},
+			"0600",
+		},
+		{
+			"with relative file path without permission",
+			map[string]string{
+				"vault.hashicorp.com/agent-inject-secret-foo": "db/creds/foo",
+				"vault.hashicorp.com/agent-inject-file-foo":   "nested/foofile",
+			},
+			"",
+		},
+		{
+			"with absolute file path",
+			map[string]string{
+				"vault.hashicorp.com/agent-inject-secret-foo": "db/creds/foo",
+				"vault.hashicorp.com/agent-inject-file-foo":   "/special/volume/foofile",
+				"vault.hashicorp.com/agent-inject-perms-foo":  "0600",
+			},
+			"0600",
+		},
+		{
+			"with absolute file path without permission",
+			map[string]string{
+				"vault.hashicorp.com/agent-inject-secret-foo": "db/creds/foo",
+				"vault.hashicorp.com/agent-inject-file-foo":   "/special/volume/foofile",
+			},
+			"",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			pod := testPod(tt.annotations)
+			var patches []*jsonpatch.JsonPatchOperation
+
+			agentConfig := basicAgentConfig()
+			err := Init(pod, agentConfig)
+			if err != nil {
+				t.Errorf("got error initialising pod, shouldn't have: %s", err)
+			}
+
+			agent, err := New(pod, patches)
+			if err != nil {
+				t.Errorf("got error creating agent, shouldn't have: %s", err)
+			}
+			cfg, err := agent.newConfig(true)
+			if err != nil {
+				t.Errorf("got error creating Vault config, shouldn't have: %s", err)
+			}
+
+			config := &Config{}
+			if err := json.Unmarshal(cfg, config); err != nil {
+				t.Errorf("got error unmarshalling Vault config, shouldn't have: %s", err)
+			}
+			if config.Templates[0].Perms != tt.permission {
+				t.Errorf("wrong permission: %s != %s", config.Templates[0].Perms, tt.permission)
 			}
 		})
 	}
@@ -468,7 +560,14 @@ func TestConfigVaultAgentTemplateConfig(t *testing.T) {
 			&TemplateConfig{ExitOnRetryFailure: false},
 		},
 		{
-			"exit_on_retry_failure absent",
+			"static_secret_render_interval 10s",
+			map[string]string{
+				AnnotationTemplateConfigStaticSecretRenderInterval: "10s",
+			},
+			&TemplateConfig{ExitOnRetryFailure: true, StaticSecretRenderInterval: "10s"},
+		},
+		{
+			"template_config_empty",
 			map[string]string{},
 			&TemplateConfig{ExitOnRetryFailure: true},
 		},