+++ subcommand/injector/flags.go
@@ -34,6 +34,10 @@ type Specification struct {
 	// AGENT_INJECT_TEMPLATE_CONFIG_EXIT_ON_RETRY_FAILURE environment variable.
 	TemplateConfigExitOnRetryFailure string `split_words:"true"`
 
+	// TemplateConfigStaticSecretRenderInterval is the
+	// AGENT_INJECT_TEMPLATE_STATIC_SECRET_RENDER_INTERVAL environment variable.
+	TemplateConfigStaticSecretRenderInterval string `envconfig:"AGENT_INJECT_TEMPLATE_STATIC_SECRET_RENDER_INTERVAL"`
+
 	// TLSAuto is the AGENT_INJECT_TLS_AUTO environment variable.
 	TLSAuto string `envconfig:"tls_auto"`
 
@@ -107,6 +111,8 @@ func (c *Command) init() {
 		`Supported log formats: "standard", "json".`)
 	c.flagSet.BoolVar(&c.flagExitOnRetryFailure, "template-config-exit-on-retry-failure", agent.DefaultTemplateConfigExitOnRetryFailure,
 		fmt.Sprintf("Value for Agent's template_config.exit_on_retry_failure. Defaults to %t.", agent.DefaultTemplateConfigExitOnRetryFailure))
+	c.flagSet.StringVar(&c.flagStaticSecretRenderInterval, "template-static-secret-render-interval", "",
+		"Value for Agent's template_config.exit_on_retry_failure.")
 	c.flagSet.StringVar(&c.flagAutoName, "tls-auto", "",
 		"MutatingWebhookConfiguration name. If specified, will auto generate cert bundle.")
 	c.flagSet.StringVar(&c.flagAutoHosts, "tls-auto-hosts", "",
@@ -203,6 +209,10 @@ func (c *Command) parseEnvs() error {
 		if err != nil {
 			return err
 		}
+	}
+
+	if envs.TemplateConfigStaticSecretRenderInterval != "" {
+		c.flagStaticSecretRenderInterval = envs.TemplateConfigStaticSecretRenderInterval
 	}
 
 	if envs.TLSAuto != "" {